fix:modify disc alarm

+++ core/broadcast/common/src/broadcast_dfx_event.c
@@ -25,7 +25,7 @@ void BroadcastDiscEvent(int32_t eventScene, int32_t eventStage, DiscEventExtra *
     }
  
     DiscEventExtra extra = { 0 };
-    uint64_t stamptime = SoftBusGetSysTimeMs();
+    int32_t stamptime = (int32_t)SoftBusGetSysTimeMs();
     for (int32_t i = 0; i < size; i++) {
         extra.capabilityBit = discExtra[i].capabilityBit;
         extra.discType = discExtra[i].discType;
